% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-ticks.R
\name{add_ticks}
\alias{add_ticks}
\title{Add labelled and unlabelled tick marks to a plot, for linear or log axes
(where labels represent unlogged values),
calculated automatically if needed, or can be explicitly specified.}
\usage{
add_ticks(
  log = "",
  tcl_small = -0.2,
  mgp_val = c(1.6, 0.5, 0),
  x_big_ticks = NULL,
  x_big_ticks_labels = NULL,
  x_small_ticks = NULL,
  x_small_ticks_by = NULL,
  x_small_ticks_labels = NULL,
  y_big_ticks = NULL,
  y_big_ticks_labels = NULL,
  y_small_ticks = NULL,
  y_small_ticks_by = NULL,
  y_small_ticks_labels = NULL
)
}
\value{
Adds axes and big and small tick marks to the plot. Returns invisible.
}
\description{
Useful because you can then interpret the unlogged values, e.g. Figures 2(h)
and 6(b) of MEE paper and Figure 7 of MEPS paper. \code{log} argument needs to
match the original \code{\link[=plot]{plot()}} call (this is not checked).
}
\details{
Legends are automatically set, but can be tailored with the
arguments defined below.
}
\examples{
\dontrun{
plot(1:10, axes = FALSE)
add_ticks()
box()

plot(1:1000, axes = FALSE, log = "xy")
add_ticks(log = "xy")
box()
}
}
\author{
Andrew Edwards
}
