% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-bin-breaks.R
\name{calc_bin_breaks}
\alias{calc_bin_breaks}
\title{Given a tibble that contains a column of measurements, append minimum and
maximum for each bin}
\usage{
calc_bin_breaks(
  tib,
  measurement = "length",
  bin_width = 1,
  represents = "midpoint"
)
}
\arguments{
\item{tib}{tibble values including a column identified by \code{measurement}}

\item{measurement}{character description of the column of interest, to then
use to generate names for the new columns}

\item{bin_width}{numeric value of assumed bin width. If multiple bin widths
are needed then split the data up and apply the function each time, then
combine together. If multiple bin widths might have been used for one type
of data but we can't filter those easily, then see \code{\link[=calc_max_of_bin]{calc_max_of_bin()}}
TODO check.}

\item{represents}{character describing what the \code{measurement} column
is assumed to be: either the \code{min} (minimum), \code{midpoint}, or \code{max} (maximum) of the bins.}
}
\value{
tibble of the input with appended columns \verb{<measurement>_bin_min}
and \verb{<measurement>_bin_max}.
}
\description{
Given a tibble of measurements identified by \code{measurement}, create new
columns of \verb{<measurement>_bin_min} and \verb{<measurement>_bin_max}
representing minimum and maximum of each bin. Options specify whether the
original \code{measurement} represents the midpoint of minimum of each bin. Bin
width can also be specified.
}
\examples{
\dontrun{
# TODO
# calc_bin_breaks(mediterranean_data, represents = "max") # etc.
calc_max_of_bin(c(1:5, 7, 9.83, 11.24)) \%>\% a()
calc_max_of_bin(c(1:5, 7, 9.83, 11.24), bin_widths = c(0.01, 1))  \%>\% a()
calc_max_of_bin(c(1:5, 7, 9.83, 11.24), vec_represents_bin_mid = FALSE)  \%>\% a()
calc_max_of_bin(c(1:5, 7, 9.83, 11.24), vec_represents_bin_mid = FALSE, bin_widths = c(0.01, 1))  \%>\% a()

}
}
\author{
Andrew Edwards
}
