% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-xmin-and-fit-mlebin.R
\name{determine_xmin_and_fit_mlebin}
\alias{determine_xmin_and_fit_mlebin}
\title{Given data for MLEbin calculation, determine a common binning and use that
to calculate \code{x_min} and then fit using MLEbin.}
\usage{
determine_xmin_and_fit_mlebin(
  dat,
  bin_width = 1,
  bin_start = 0,
  x_min = NULL,
  ...
)
}
\arguments{
\item{dat}{tibble of data in the format required for fitting
using MLEbin method; see \code{\link[=fit_size_spectrum]{fit_size_spectrum()}}.}
}
\value{
list of class \code{determine_xmin_and_fit_mlebin} for plotting, containing TODO
see \code{\link[=determine_xmin_and_fit]{determine_xmin_and_fit()}} also
}
\description{
Given data for MLEbin calculation, determine a common binning and use that
to calculate \code{x_min} and then fit using MLEbin.
}
\examples{
\dontrun{
 TODO
}
}
\author{
Andrew Edwards
}
