% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-xmin-and-fit-mlebins.R
\name{determine_xmin_and_fit_mlebins}
\alias{determine_xmin_and_fit_mlebins}
\title{Given data for MLEbins calculation, with species-specific weight bins,
determine a common binning and use that to calculate \code{x_min} and then fit
using MLEbins. TODO might also work for just MLEbin, need to test.}
\usage{
determine_xmin_and_fit_mlebins(
  dat_for_mlebins,
  bin_width = 1,
  bin_start = 0,
  x_min = NULL,
  ...
)
}
\arguments{
\item{dat_for_mlebins}{tibble of data in the format required for fitting
using MLEbins method. TODO}
}
\value{
list of class \code{determine_xmin_and_fit_mlebins} for plotting, containing TODO
see \code{\link[=determine_xmin_and_fit]{determine_xmin_and_fit()}} also
}
\description{
Given data for MLEbins calculation, with species-specific weight bins,
determine a common binning and use that to calculate \code{x_min} and then fit
using MLEbins. TODO might also work for just MLEbin, need to test.
}
\examples{
\dontrun{
 TODO
}
}
\author{
Andrew Edwards
}
