% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots-parser.R
\name{dots_parser}
\alias{dots_parser}
\title{Allow a function to parse out the ... values into two separate functions,
without giving a warning.}
\usage{
dots_parser(FUN, ...)
}
\arguments{
\item{FUN}{function to be run with arguments from \code{...}}

\item{...}{arguments to be passed onto \code{FUN}}
}
\value{
appropriately evaluates \code{FUN}
}
\description{
Taken from
https://stackoverflow.com/questions/4124900/is-there-a-way-to-use-two-statements-in-a-function-in-r. So within a larger function that has \code{...}, use this to send the appropriate values to two (or more) further functions. The values can go to both further functions.
}
\examples{
\dontrun{
# Code within a function would have previously looked like:
func1(a = myvaluea, b = myvalueb, ...)
But that returns an error because ... includes arguments not accepted by func1, this now works:
dots_parser(func1, a = myvaluea, b = myvalueb, ...)
}
}
\author{
Mike Blazanin in the Stack Overflow post referenced above, written up here by Andrew Edwards
}
