% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-bins-to-body-mass-bins.R
\name{length_bins_to_body_mass_bins}
\alias{length_bins_to_body_mass_bins}
\title{Convert data of counts of species in length bins to counts in body-mass bins
using species-specific length-weight coefficients}
\usage{
length_bins_to_body_mass_bins(
  dat,
  coefficients,
  length_relationship_unit = "cm",
  body_mass_relationship_unit = "g",
  length_data_unit = "cm"
)
}
\arguments{
\item{dat}{data.frame where each row is a count of a particular species in a
particular length bin. It must include the columns (can have more):
\itemize{
\item \code{species} - name of the species
\item \code{length_bin_min} - minimum of the length bin
\item \code{length_bin_max} - maximum of the length bin
\item \code{bin_count} - count in the bin, can be non-integer (as can happen when
raw counts are normalised to account for sampling effort).
}}

\item{coefficients}{data.frame of species-specific length-weight
coefficients, with columns:
\itemize{
\item \code{species} - name of the species; must match the conventions used in \code{dat}
\item \code{alpha}, \code{beta} - length-weight coefficients for the given species
}}

\item{length_relationship_unit}{character giving the units used for $l$ in the the length-weight
relationship; must be mm or cm (the default)}

\item{body_mass_relationship_unit}{character giving the units used for $w$ in the
length-weight relationship; must be g (the default) or kg}

\item{length_data_unit}{character giving the units of the length data, must
be mm or cm (the default).}
}
\value{
data.frame with each row representing a body-mass bins, and columns:
\itemize{
\item \code{species} - unique species corresponding to \code{bin_min} and \code{bin_max} in
the corresponding row. If any resulting bin definitions are the same for
two or more distinct species, then this column is omitted. This could
occur if the same length-weight relationship is used for different species
(e.g. same relationship used for multiple rockfish species), or just by
chance. The \code{species} is not needed for the subsequent MLEbins method, but
it can be helpul for users to still keep track of species if they body-mass
bins each correspond to just one species.
\item \code{bin_min} - minimum of the body-mass bin (g)
\item \code{bin_max} - maximum of the body-mass bin (g)
\item \code{bin_count} - total count of individuals within that body-mass bin; can
be non-integer.
}
}
\description{
For each species $s$, the length-weight relationship is
$w = \alpha_s l^{\beta_s}$
where $w$ is the estimated body mass (g) of an individual, $l$ is the known
body length (cm), and $alpha_s$ and $\beta_s$ are species-specific
parameters; this is equation (1) in our \href{https://www.int-res.com/abstracts/meps/v636/p19-33/}{MEPS paper}.
}
\details{
Each count corresponds to numbers of a given species in a given length bin,
and we use the species-specific coefficients to convert the length bins to
body mass bins (see the MEPS paper for more details). Body-mass bins that
are repeated (likely from the same species being measured in the same length
bin, though could also occur coincidentally for different species) are
aggregated to give total counts for such bins, so that there are no
duplicated bins in the output. The resulting tibble is also sorted in
order of \code{bin_min}, starting with the lowest. Any species that do not have
length-weight coefficients get omitted from the resulting tibble.

The resulting species-specific body-mass bins can then be used for the
MLEbins method.
}
\examples{
\dontrun{
 # Example using the first 1000 lines of the Mediterranean data in the
  package, copy from vignette:
med_subset <- mediterranean_data[1:1000, ]
# TODO need to functionalise the conversion of lengths to length bins
 mediterranean_example <-
  length_bins_to_body_mass_bins(med_subset,
                                mediterranean_length_weight_coefficients,
TODO, ...)
}
}
\author{
Andrew Edwards
}
