% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-hist-for-binned-counts-mlebin.R
\name{make_hist_for_binned_counts_mlebin}
\alias{make_hist_for_binned_counts_mlebin}
\title{Take a data.frame to be used for the MLEbin method and create the histogram
object to be used to determine \code{x_min}}
\usage{
make_hist_for_binned_counts_mlebin(dat)
}
\arguments{
\item{dat}{tibble of data in the format required for fitting
using MLEbin or MLEbins methods; i.e. at a minimum has to include the columns:
* \code{bin_min}
* \code{bin_max}
* \code{bin_count}.}
}
\value{
a histogram list object with components (see \code{?hist}):
\itemize{
\item \code{breaks}
\item \code{mids}
\item \code{counts}
\item \code{xname}
\item \code{equidist} will be TRUE or FALSE depending on whether the bins in \code{dat}
have equal bin widths
}
}
\description{
Called from \code{\link[=determine_xmin_and_fit_mlebin]{determine_xmin_and_fit_mlebin()}}, adapting from
\code{\link[=make_hist_for_binned_counts]{make_hist_for_binned_counts()}}. This version should be simpler, as we are
not making up new bins (had to for MLEbins data because the bins overlap),
just using those specified by the user.
}
\details{
Gives values as a histogram list object.
Can then use \code{plot()} which calls \code{plot.histogram()}.
}
\examples{
\dontrun{
TODO counts_per_bin_example
make_hist(counts_per_bin_example)
TODO for test do
}
}
\author{
Andrew Edwards
}
