% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negll-mle.R
\name{negll_mle}
\alias{negll_mle}
\title{Calculate negative log-likelihood for the bounded power-law
distribution}
\usage{
negll_mle(b, x, n, x_min, x_max, sum_log_x)
}
\arguments{
\item{b}{value of \code{b} for which to calculate the negative log-likelihood}

\item{x}{vector of values of data (e.g. masses of individual fish)}

\item{n}{\code{length(x)}, have as an input to avoid repeatedly calculating it when
function is called multiple times in an optimization routine}

\item{x_min}{minimum value of \code{x} to avoid repeatedly calculating}

\item{x_max}{maximum value of \code{x} to avoid repeatedly calculating}

\item{sum_log_x}{\code{sum(log(x))} to avoid repeatedly calculating}
}
\value{
numeric negative log-likelihood of the parameters given the data
}
\description{
Calculate the negative log-likelihood of the parameters \code{b}, \code{x_min} and
\code{x_max} given data \code{x} for the PLB model. Returns the negative
log-likelihood. Will be called by \code{nlm()} or similar. \code{x_min} and \code{x_max}
are just estimated as the min and max of the data, not numerically using likelihood.
}
\author{
Andrew Edwards
}
