% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-biomass.R
\name{p_biomass}
\alias{p_biomass}
\title{Biomass cumulative distribution function from MEE equations A.4 and A.8}
\usage{
p_biomass(x, b, xmin, xmax, n)
}
\arguments{
\item{x}{vector of values for which to calculate the total biomass between
\code{xmin} and the value}

\item{b}{estimated exponent of the PLB distribution}

\item{xmin}{minimum bound of the distribution, \code{xmin > 0}}

\item{xmax}{maximum bound for bounded distribution, \code{xmax > xmin}}

\item{n}{number of individuals (or total counts)}
}
\value{
return vector of total biomass between \code{xmin} and each value of \code{x}
}
\description{
Total biomass between \code{xmin} and \code{x}, assuming a bounded power-law
distribution of body masses between \code{xmin} and \code{xmax} and a given value of
exponent \code{b}, and a total of \code{n} individuals.
}
\details{
Given by MEE equations A.4 and A.8. Can then be called by \code{\link[=p_biomass_bins]{p_biomass_bins()}}
to give total biomass (and normalised biomass) in each bin.
}
\examples{
\dontrun{
p_biomass(x = c(1, 5, 10, 20, 50, 100),
          b = -2,
          xmin = 1,
          xmax = 100,
          n = 1000)
}
}
\author{
Andrew Edwards
}
