% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLB.R
\name{PLB}
\alias{PLB}
\alias{dPL}
\alias{pPL}
\alias{rPL}
\alias{dPLB}
\alias{pPLB}
\alias{rPLB}
\alias{qPLB}
\title{Probability density and distribution functions, and random number
generation, for the bounded power-law distribution.}
\usage{
dPL(x = 1, b = -2, xmin = 1)

pPL(x = 10, b = -2, xmin = 1)

rPL(n = 1, b = -2, xmin = 1)

dPLB(x = 1, b = -2, xmin = 1, xmax = 100)

pPLB(x = 10, b = -2, xmin = 1, xmax = 100)

rPLB(n = 1, b = -2, xmin = 1, xmax = 100)

qPLB(p = 0.1, b = -2, xmin = 1, xmax = 100)
}
\arguments{
\item{x}{vector of values to compute the density and distribution functions.}

\item{b}{exponent of the distribution (must be <-1 for \code{\link[=PL]{PL()}})}

\item{xmin}{minimum bound of the distribution, \code{xmin > 0}}

\item{n}{number of random numbers to be generated (if \code{length(n) > 1} then
generate \code{length(n)} values)}

\item{xmax}{maximum bound for bounded distribution, \code{xmax > xmin}}

\item{p}{vector of probabilities for \code{qPLB()}}
}
\value{
\code{\link[=dPLB]{dPLB()}} returns vector of probability density values
corresponding to \code{x}. \code{\link[=pPLB]{pPLB()}} returns vector of cumulative
distribution values P(X <= x) corresponding to \code{x}. \code{\link[=rPLB]{rPLB()}} returns
a vector (of length \code{n}) of independent random draws from the distribution.
\code{\link[=qPLB]{qPLB()}} returns vector of values of \code{x} for which \code{P(X <= x) = p} (each
element corresponding to the element of \code{p}). So
\code{pPLB(qPLB(seq(0, 1, by = 0.1)))} gives \verb{0, 0.1, ..., 1}.
}
\description{
For the bounded power-law distribution (PLB) functions in the standard R
format (like \code{\link[=dnorm]{dnorm()}} etc.) for density, distribution, and random number generation.
These are:
\itemize{
\item probability density function, \code{\link[=dPLB]{dPLB()}}
\item cumulative distribution function P(X <= x), \code{\link[=pPLB]{pPLB()}}
\item random generation of values \code{\link[=rPLB]{rPLB()}},
as described in Edwards et al. (2017, Methods in Ecology and
Evolution, 8:57-67). Random generation uses the inverse method (e.g. p1215 of Edwards
2008, Journal of Animal Ecology, 77:1212-1222). \code{\link[=qPLB]{qPLB()}} is from inverting the
cumulative distribution function. Unbounded distributions
are also included for completeness (except for qPL) but are likely not needed.
}
}
\examples{
\dontrun{
rPLB(10)
rPLB(10, b = -2.5)
dPLB(1:10, xmin = 2)
pPLB(1:10, xmin = 2)
qPLB(c(0.05, 0.5, 0.95))
}
}
\author{
Andrew Edwards
}
