% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.determine_xmin_and_fit.R
\name{plot.determine_xmin_and_fit}
\alias{plot.determine_xmin_and_fit}
\title{Plot results from determining xmin by mode method and then fitting use MLE hake spectra results}
\usage{
\method{plot}{determine_xmin_and_fit}(res, xlim_hist = NULL, ...)
}
\arguments{
\item{res}{list of class \code{determine_xmin_and_fit} as output from \code{determine_xmin_and_fit()}}

\item{xlim_hist}{numeric vector of two values representing \code{xlim} for histogram plot; default is the full range
of breaks (which might be too large for a clear figure, especially given
the linear axis). TODO could maybe add in log-x axis, but not option in
\code{plot.hist()} so would have to make it up here.}

\item{...}{arguments to passed onto \code{\link[=hist]{hist()}} for now, but probably want
more. TODO \code{sizeSpectra::ISD_bin_plot_nonoverlapping()}}
}
\value{
figure in current device
}
\description{
TODO check if want features from plot.determine_xmin_and_fit_mlebins(),
which was adapted and improved from this.
}
\examples{
\dontrun{
See .Rmd
}
}
\author{
Andrew Edwards
}
