% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.size_spectrum_mlebin.R
\name{plot.size_spectrum_mlebin}
\alias{plot.size_spectrum_mlebin}
\title{Plot binned ISD plots similar to MEPS Figure 7 (but with nonoverlapping
bins).}
\usage{
\method{plot}{size_spectrum_mlebin}(
  res_mlebin,
  style = "log_y_axis",
  xlim = c(min(res_mlebin$data$bin_min), max(res_mlebin$data$bin_max)),
  ylim = NA,
  x_plb = NA,
  inset_label = c(0, -0.02),
  xlab = expression(paste("Body mass, ", italic(x), "(g)")),
  y_scaling = 0.75,
  mle_round = 2,
  legend_label_a = "(a)",
  legend_label_b = "(b)",
  legend_label_single = NULL,
  legend_text_a = paste0("b=", round(res_mlebin$b_mle, mle_round)),
  legend_text_a_n = paste0("n=", round(sum(res_mlebin$data$bin_count))),
  legend_text_b = NULL,
  legend_text_b_n = NULL,
  seg_col = "black",
  par_mai = c(0.4, 0.5, 0.05, 0.3),
  par_cex = 0.7,
  ...
)
}
\arguments{
\item{res_mlebin}{size_spectrum_mlebin object resulting from running
\code{fit_size_spectrum()} on binned data (such that the function
\code{fit_size_spectrum_mlebin()} is used; see vignette TODO). # TODO link help
to plot.size_spectrum_numeric - may want to put them all together}

\item{style}{character either:
\itemize{
\item \code{"log_y_axis"} - single ISD plot with logarithmic y axis (Fig. 6b of MEE paper)
\item \code{"linear_y_axis"} - for single ISD plot with linear y axis
\item \code{"both_y_axes"} - both the above plots as a two-panel plot
\item \code{"biomass_and_isd"} - to use only if the data represent body masses. Does two-panel
plot, essentially the recommended Fig. 6 of MEE paper where the top panel
is bins of normalized biomass (but improved here by showing bins in the top
panel rather than points) and the \code{"log_y_axis"} plot described above.
TODO prob have to make that consistent in other plotting function
}}

\item{x_plb}{vector of values to use to plot the fitted PLB curve; if NA then
automatically calculated (sometimes need to manually extend it to hit the
x-axis, but tricky to automate that on a log-scale)}

\item{xlab}{ylab x/y labels, explicitly given default values here which can
be modified as required.}

\item{y_scaling}{numeric scaling of y-minimum of y-axis. Axis can't go to zero on
log-log plot, but goes to the proportion \code{y_scaling} (<1)
of the minimum value of counts greater than the highest \code{bin_min} value. Do
such that can see the right-most bin in all plots.}

\item{legend_label_a}{character label (default \code{"(a)"}) to use for panel a for two-panel plot
(\code{log_y_axis = "both"}).}

\item{legend_label_b}{character label to use for panel b for two-panel plot
(\code{log_y_axis = "both"}).}

\item{legend_label_single}{character label to use for the only panel for a one-panel plot (\code{log_y_axis = "yes"} or \code{"no"}).}

\item{legend_text_a}{text to include in the legend for panel
a for two-panel plot (\code{log_y_axis = "both"}), the \code{b = -1.58} in Fig. 7a
of MEPS paper, or the only panel for a one-panel plot (\code{log_y_axis = "yes"} or \code{"no"}).}

\item{legend_text_a_n, }{legend_text_b_n as for \code{legend_text_a} and
\code{legend_text_b} but for another row of information, default being \verb{n = <sample size>} as in Fig. 7a of MEPS paper.}

\item{legend_text_b}{text to include in the legend for panel
b for two-panel plot (\code{log_y_axis = "both"}); ignored for one-panel plot}

\item{seg_col}{colour to use for the segments (green in Fig. 7 of
MEPS paper).}

\item{par_mai}{vector of values to use for \code{par(mai)}}

\item{par_cex}{numeric value to use for \code{par(cex)} (font size)}

\item{...}{Further arguments for \code{plot_isd()} and then \code{plot()}, except
cannot have \code{log} as that gets overridden. TODO check}

\item{log_y_axis}{TODO delete a bunch of these params.....character either \code{"both"}, do two plots (like Fig. 7 of
MEPS paper), \code{"no"} for single plot with linear y axis (like Fig. 7a of
MEPS paper), \code{"yes"} for single plot with logarithmic y axis (like Fig. 7b
of MEPS paper). Legends are automatically set, but can be tailored with the
arguments defined below.}
}
\value{
one- or two-panel plot of the ISD with data in binned form like in
Fig. 7, 7a or 7b (depending on settings) of MEPS paper, but with nonoverlapping bins; returns nothing.
}
\description{
TODO need xlab here and called functions
}
\details{
Plots one- or two-panel plot of the ISD with data in binned form like in
Fig. 7, 7a or 7b (depending on settings) of MEPS paper, but with
nonoverlapping bins. See \code{log_y_axis} to specify exact plot(s).

Also directly called from \code{plot.size_spectrum_mlebins()} with no
extra arugments for MLEbins method.
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
