% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.size_spectrum_mlebins.R
\name{plot.size_spectrum_mlebins}
\alias{plot.size_spectrum_mlebins}
\title{Plot a binned ISD plots similar to MEPS Figure 7 from MLEbins method.}
\usage{
\method{plot}{size_spectrum_mlebins}(res_mlebins, ...)
}
\arguments{
\item{...}{Further arguments for \code{plot_isd()} and then \code{plot()} TODO check}
}
\value{
one- or two-panel plot of the ISD with data in binned form like in
Fig. 7, 7a or 7b (depending on settings) of MEPS paper, with overlapping bins; returns nothing.
}
\description{
Think can just pass everything on, need to test TODO. And just link to
plot.size_spectrum_mlebin() help.
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
