% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.size_spectrum_numeric.R
\name{plot.size_spectrum_numeric}
\alias{plot.size_spectrum_numeric}
\title{Plot individual size distribution of values and the MLE fit with
confidence intervals}
\usage{
\method{plot}{size_spectrum_numeric}(
  res,
  style = "log_y_axis",
  xlim = c(min(res$x), max(res$x)),
  ylim = NULL,
  x_plb = NULL,
  y_scaling = 0.75,
  mle_round = 2,
  inset_label = c(0, -0.02),
  xlab = expression(paste("Body mass, ", italic(x), "(g)")),
  legend_label_a = "(a)",
  legend_label_b = "(b)",
  legend_label_single = NULL,
  legend_text_a = paste0("b=", round(res$b_mle, mle_round)),
  legend_text_a_n = paste0("n=", round(length(res$x))),
  legend_text_b = NULL,
  legend_text_b_n = NULL,
  par_mai = c(0.4, 0.5, 0.05, 0.3),
  par_cex = 0.7,
  ...
)
}
\arguments{
\item{res}{size_spectrum_numeric object, as output from
\code{\link[=fit_size_spectrum.numeric]{fit_size_spectrum.numeric()}}, which gets called when applying
\code{\link[=fit_size_spectrum]{fit_size_spectrum()}} to a numeric vector}

\item{style}{character either:
\itemize{
\item \code{"log_y_axis"} - single ISD plot with logarithmic y axis (Fig. 6b of MEE paper)
\item \code{"linear_y_axis"} - for single ISD plot with linear y axis
\item \code{"both_y_axes"} - both the above plots as a two-panel plot
\item \code{"biomass_and_isd"} - to use only if the data represent body masses. Does two-panel
plot, essentially the recommended Fig. 6 of MEE paper where the top panel
is bins of normalized biomass (but improved here by showing bins in the top
panel rather than points) and the \code{"log_y_axis"} plot described above.
TODO prob have to make that consistent in other plotting function
}}

\item{xlab}{ylab x/y labels, explicitly given default values here which can
be modified as required.}

\item{...}{Further arguments for \code{plot_isd()} and then \code{plot()}, except
cannot have \code{log} as that gets overridden. TODO check}
}
\value{
One- or two-panel plot of raw data and PLB distribution (and fits of
confidence limits) as solid (and dashed) fitted using MLE method; returns
nothing. TODO could return invisible biomass calcs
}
\description{
Plots one- or two-panel plot of the ISD and data, with maximum likelihood
fits shown, as calcualted using the MLE method.. Two-panel plot is either
the ISD style (individual points and fitted distribution) with (a) linear
y-axis and (b) logarithimc y-axis (so like Fig 7 of MEPS paper but for
unbinned data), or, only for when
the data represent body masses, (a) the normalised biomass on log-log
axes with fitted estimates and (b) same as (b) above, essentially the
recommended Fig. 6 of the MEE paper, but improved by showing bins in the top
panel rather than points.
Single plots are either the ISD style with logarithmic or linear y-axis. See
the \code{style} argument.
}
\details{
Legends are automatically set, but can be tailored with the
arguments defined below.
}
\examples{
\dontrun{
# TODO
res_vec <- fit_size_spectrum(sim_vec)
plot(res_vec)
plot(res_vec, log = "x")
plot(res_vec, log = "")
plot(res_vec, x_small_ticks_labels = c(5, 50, 500), log = "x") # Tailor the
  labels for a particular figure
}
}
\author{
Andrew Edwards
}
