% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-isd-binned.R
\name{plot_isd_binned}
\alias{plot_isd_binned}
\title{Plot a single binned ISD plot, as called from \code{plot.size_spectrum_mlebin()},
or used directly from \code{plot.size_spectrum_numeric()} to make an LBN-style
plot of individual values and MLE plot.}
\usage{
plot_isd_binned(
  res_mlebin,
  log,
  xlim,
  ylim,
  x_plb,
  y_plb,
  y_plb_conf_min,
  y_plb_conf_max,
  plot_conf_ints = TRUE,
  xlab = expression(paste("Values, ", italic(x))),
  ylab = expression(paste("Total ", counts >= x), sep = ""),
  mgp_val = c(1.6, 0.5, 0),
  tcl_small = -0.2,
  inset_label = c(0, 0),
  inset_text = c(0, 0.04),
  legend_label = NULL,
  legend_text = NULL,
  legend_text_n = NULL,
  legend_text_second_row_multiplier = 2,
  legend_position = "topright",
  x_big_ticks = NULL,
  x_big_ticks_labels = NULL,
  x_small_ticks = NULL,
  x_small_ticks_by = NULL,
  x_small_ticks_labels = NULL,
  y_big_ticks = NULL,
  y_big_ticks_labels = NULL,
  y_small_ticks = NULL,
  y_small_ticks_by = NULL,
  y_small_ticks_labels = NULL,
  y_scaling = 0.75,
  seg_col = "green",
  rect_col = "grey",
  fit_col = "red",
  fit_lwd = 2,
  conf_lty = 2,
  show_fit_on_top = TRUE
)
}
\arguments{
\item{res_mlebin}{size_spectrum_mlebin object resulting from running
\code{fit_size_spectrum()} on binned data (such that the function
\code{fit_size_spectrum_mlebin()} is used; see vignette TODO). # TODO link help
to plot.size_spectrum_numeric - may want to put them all together}

\item{x_plb}{vector of values to use to plot the fitted PLB curve; if NA then
automatically calculated (sometimes need to manually extend it to hit the
x-axis, but tricky to automate that on a log-scale)}

\item{y_plb}{vector of values corresponding to the MLE fit of the PLB
distribution at each value of \code{x_PLB}}

\item{y_plb_conf_min, }{y_plb_conf_max vector of values corresponding to the fit, using the
minimum/maximum value confidence interval for exponent b, of the PLB
distribution at each value of \code{x_PLB}}

\item{plot_conf_ints}{logical whether to plot confidence intervals or not}

\item{xlab}{ylab x/y labels, explicitly given default values here which can
be modified as required.}

\item{legend_text_second_row_multiplier}{numeric multiplier of the second
row of legend text to space it out, especially for smaller panel plots.}

\item{y_scaling}{numeric scaling of y-minimum of y-axis. Axis can't go to zero on
log-log plot, but goes to the proportion \code{y_scaling} (<1)
of the minimum value of counts greater than the highest \code{bin_min} value. Do
such that can see the right-most bin in all plots.}

\item{seg_col}{colour to use for the segments (green in Fig. 7 of
MEPS paper).}

\item{fit_col}{colour to use for fitted curves}

\item{fit_lwd}{line thickness to use for fitted curves}

\item{conf_lty}{line type to use for confidence intervals}

\item{show_fit_on_top}{logical, whether to plot the fitted PLB curve on top of the
data or underneath. Usually on top (the default) is fine, but sometimes
having it underneath is better.}

\item{par_mai}{vector of values to use for \code{par(mai)}}

\item{par_cex}{numeric value to use for \code{par(cex)} (font size)}

\item{LBN_style}{Whether to plot an LBN style plot (similar to Fig. 6a of MEE
paper), used directly from \code{plot.size_spectrum_numeric()} to give the
recommended Fig. 6 plot for MLE results, provided data are body masses.}
}
\value{
single panel plot of the ISD with data in binned form like in
Fig. 7a or 7b of MEPS paper, but with nonoverlapping bins; returns nothing.
}
\description{
Plot a single binned ISD plot, as called from \code{plot.size_spectrum_mlebin()},
or used directly from \code{plot.size_spectrum_numeric()} to make an LBN-style
plot of individual values and MLE plot.
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
