% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lbn-style.R
\name{plot_lbn_style}
\alias{plot_lbn_style}
\title{Biomass size spectrum plot on log-log scale for body-mass data demonstrating
uncertainties}
\usage{
plot_lbn_style(
  res,
  x_plb,
  xlim = NULL,
  ylim = NULL,
  plot_conf_ints = TRUE,
  plot_binned_fitted = TRUE,
  xlab = NULL,
  ylab = "Normalised biomass",
  mgp_val = c(1.6, 0.5, 0),
  tcl_small = -0.2,
  inset_label = c(0, 0),
  inset_text = c(0, 0.04),
  legend_label = NULL,
  legend_text = NULL,
  legend_text_n = NULL,
  legend_text_second_row_multiplier = 2,
  legend_position = "topright",
  x_big_ticks = NULL,
  x_big_ticks_labels = NULL,
  x_small_ticks = NULL,
  x_small_ticks_by = NULL,
  x_small_ticks_labels = NULL,
  y_big_ticks = NULL,
  y_big_ticks_labels = NULL,
  y_small_ticks = NULL,
  y_small_ticks_by = NULL,
  y_small_ticks_labels = NULL,
  y_scaling = 0.75,
  seg_col = "green",
  rect_col = "grey",
  fit_col = "red",
  fit_lwd = 2,
  conf_lty = 2,
  ...
)
}
\arguments{
\item{res}{results of class \code{size_spectrum_numeric} or
\code{size_spectrum_mlebin}
ignore I think: TODO but if called
from \code{plot.size_spectrum_numeric()} then the results are from the MLE
method and do not have all the columns that would have from MLEbin method.

TODO onwards:}

\item{...}{further arguments to be passed to \code{plot()} and
\code{plot_binned_fitted()}}

\item{binValsTibble}{tibble of binned data with each row representing a bin
and with columns \code{binMin} and \code{binMmax} (min and max break of each bin)
and \code{binCount} (count in that bin), as in the \code{binVals} component of the
output of \code{binData}. \code{wmin} and \code{wmax} can also be used instead of
\code{binMin} and \code{binMax}. Extra columns are ignored.}

\item{binBreaks}{vector of bin breaks}

\item{binCounts}{vector of bin counts}

\item{b.MLE}{maximum likelihood estimate of \emph{b} (ideally from the MLEbin method)}

\item{b.confMin}{lower 95\\% confidence limits of \emph{b}}

\item{b.confMax}{upper 95\\% confidence limits of \emph{b}}

\item{plot.binned.fitted}{if TRUE then also plot the estimated normalised
biomass in each bin for the MLE of \emph{b} and it's confidence limits}

\item{log.xy}{Which axes to log, for \code{plot(..., log = log.xy)}. So "xy" for
log-log axes, "x" for only x-axis logged, "" for both axes unlogged.}

\item{x.PLB}{vector of values to use to plot the fitted PLB curve; if NA then
automatically calculated}

\item{legend}{if TRUE then add legend}

\item{leg.pos}{position of legend, from "bottomright"', '"bottom"',
'"bottomleft"', '"left"', '"topleft"', '"top"', '"topright"', '"right"'
and '"center"'.}

\item{inset}{inset distance vector for legend}

\item{leg.text}{text for legend}
}
\value{
TODO should return a tibble of results
}
\description{
Biomass size spectrum plot for binned data on a log-log scale like people
have been used to seeing. Shows the bin widths
explicitly and the normalised biomass in
each bin (with resulting uncertainties if applicable). So extending MEE Fig. 6 for already
binned data, using a new approach motivated by MEPS Fig. 7. Should probably
then be recommended to replace MEE Fig. 6. Axes are logged, and labelled as either
logged or (more intuitive) logged values. So it's an LBN-style plot but the
fitting is done using likelihood.
}
\details{
If from the MLE method based on known individual body masses then there is
no uncertainty of the normalised biomass within each bin (no vertical
uncertainty on the plot). But if from the MLEbin method where the data are
already binned, then there is uncertainty and so we have uncertainty in the
vertical direction on the plot. Only appropriate for nonoverlapping bins.

Adapting from \code{sizeSpectra::LBN_bin_plot()} and \code{plot_isd_binned()} (wanting
it consistent with the latter, but seemed better to make a new function).

TODO use sizeSpectra vignette
\code{MLEbin_recommend.Rmd} for new vignette.
}
\author{
Andrew Edwards
}
