% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quevedo_table_b1}
\alias{quevedo_table_b1}
\title{Results of the fits as shown in Table B.1 (to use as the example code to
plot Figure B.20 of Quevedo et al., 2026)}
\format{
tibble
}
\source{
Generated from \code{data-raw/mediterranean-results.R}.
}
\usage{
quevedo_table_b1
}
\description{
Results of the fits as shown in Table B.1 (to use as the example code to
plot Figure B.20 of Quevedo et al., 2026)
}
\examples{
\dontrun{
quevedo_table_b1
knitr::kable(quevedo_table_b1,
             digits = 2)       # As shown in Table B.1 (2 decimal places)
plot_multiple_exponents(quevedo_table_b1,         # Plot Figure B.20
                        shade_first = TRUE)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
