% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers (large values with gaps from remaining continuous values)
from MLEbins (TODO others to come) results object}
\usage{
remove_outliers(dat, ...)
}
\arguments{
\item{res}{One of:
\itemize{
\item \code{sizespectrum_mlebins} object.
}}

\item{number}{numeric value for how many of the top measurements to remove;
calculate it manually for now}
}
\value{
list containing two tibbles plus two numerics. Each tibble contains just the data values needed for
calculations, which are \code{species}, \code{bin_min}, \code{bin_max}, and
\code{bin_count}. \code{count_gte_bin_min} etc. will be recalculated in the new
analysis. They are:
\itemize{
\item \code{dat-keep} records that are being kept
\item \code{dat-removed} records that are removed
Numeric values are:
\item \code{bin_count_removed} sum of the counts in bins that have been removed
\item \code{bin_count_removed_prop} proportion of the counts in bins that have been
removed (total counts removed divided by total counts in original data set).
}
}
\description{
Remove outliers (large values with gaps from remaining continuous values)
from MLEbins (TODO others to come) results object
}
\examples{
\dontrun{
remove_outliers(**)
}
}
\author{
Andrew Edwards
}
