% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_vec}
\alias{sim_vec}
\title{Simulated vector of 1000 simulated values from a power-law distribution}
\format{
numeric
}
\source{
Generated from \code{data-raw/sim_vec.R}.
}
\usage{
sim_vec
}
\description{
As used for Figures 1 and 2 of MEE paper, used as a default set of
values. Paramaters are \code{b = -2}, \code{xmin = 1}, xmax = 1000`, and the seed was
set to 42.
}
\examples{
\dontrun{
head(sim_vec)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
