% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mle-table.R,
%   R/summary-table.determine_xmin_and_fit.R
\name{summary_mle_table}
\alias{summary_mle_table}
\alias{summary_mle_table.determine_xmin_and_fit}
\title{TODO Print a Markdown summary table of main results from...  adapting from HDIanalysis}
\usage{
summary_mle_table(obj, ...)

\method{summary_mle_table}{determine_xmin_and_fit}(res, dig = 2)
}
\arguments{
\item{obj}{result from ..., either with class
\code{intervals_density} if  was applied to a vector, or of
class \code{intervals_density_list} if ...  was applied to a data
frame; the respective function ...  or
gets automatically used.}

\item{dig}{number of decimal places to show}
}
\value{
Markdown code (for automatic use straight in an R Markdown document, for instance)
}
\description{
TODO Works on MLE/MLEbin output applied to either a vector or a data
frame, creating an appropriate table that renders in an R Markdown
document. See the vignettes.
}
\examples{
\dontrun{
# TODO And see vignettes
res <- create_intervals(rec_2021)
summary_table(res)

res_df <- create_intervals(hake_recruitment_mcmc)
summary_table(res_df)
}
}
\author{
Andrew Edwards
}
