% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-confidence-interval.R
\name{calc_confidence_interval}
\alias{calc_confidence_interval}
\title{Calculate confidence interval}
\usage{
calc_confidence_interval(min_ll, x, n, x_min, x_max, sum_log_x, b_vec)
}
\arguments{
\item{min_ll}{list output from \code{\link[=nlm]{nlm()}}.}

\item{x_min}{minimum value of data to fit the PLB distribution to. If \code{NULL}
(the default) then it is set to the minimum value of the data (if \code{dat} is
\code{numeric}), else to the minimum bin break of the lowest bin or ??any
other options? Mention species-specific}

\item{x_max}{maximum value of data to fit the PLB distribution to. If \code{NULL}
(the default) then it is set to the maximum value of the data (if \code{dat} is
\code{numeric}), else to the maximum bin break of the highest bin or ??any
other options? Mention species-specific...}
}
\description{
Gets called from \code{\link[=fit_size_spectrum.numeric]{fit_size_spectrum.numeric()}}.
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
