% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-max-of_bin.R
\name{calc_max_of_bin}
\alias{calc_max_of_bin}
\title{Calculate the maximum value of a bin given a value (that can represent the
minimum or midpoint)}
\usage{
calc_max_of_bin(vec, bin_widths = 1, vec_represents_bin_min = TRUE)
}
\arguments{
\item{vec}{vector of values corresponding to bins}

\item{bin_widths}{numeric of length 1 for all bins having same widths, or
vector of two different allowed bin widths TODO think if need more, or
generalise. Or tell people to do manually if not simply two options.}
}
\value{
tibble of bin_min, bin_max
}
\description{
Also depends on binbreak.
}
\author{
Andrew Edwards
}
