% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_xmin_and_fit.R
\name{determine_xmin_and_fit}
\alias{determine_xmin_and_fit}
\title{Determine x_min as the mode and then fit it. TODO only on a numeric vector
for now}
\usage{
determine_xmin_and_fit(x, bin_width = 1, bin_start = NULL, ...)
}
\arguments{
\item{x}{\code{numeric} vector of values (such as individual body masses or lengths), which uses
the MLE method (via the function \code{\link[=fit_size_spectrum.numeric]{fit_size_spectrum.numeric()}} TODO could
make \code{dat} for consistency with fit_size_spectrum() if going to make this
more general. TODO join help files together}

\item{...}{arguments to pass onto fit....}
}
\value{
list of class \code{determine_xmin_and_fit} for plotting, containing TODO
\describe{
\item{bin_width}
\item{xmin}
\item{xmax}
\item{n}
\item{counts_per_bin}
\item{counts_per_bin_desc}{counts in the descending limb, including the
peak}
\item{b_l}
\item{b_l_confMin}
\item{b_l_confMax}
}
}
\description{
Determine x_min as the mode and then fit it. TODO only on a numeric vector
for now
}
\examples{
\dontrun{
# see .Rmd file
TODO  res_all_strata <- fit_all_years(raw_simp_prop,
                    bin_width_each_year = bin_width_each_year)
}
}
\author{
Andrew Edwards
}
