% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-hist-for-binned-counts.R
\name{make_hist_for_binned_counts}
\alias{make_hist_for_binned_counts}
\title{Take a tibble with \code{bin_min}, \code{bin_max}, and \code{bin_count} and assign counts
to new equal bins based on \code{bin_min}, summing the original counts (which can
be non-integer)}
\usage{
make_hist_for_binned_counts(dat_for_mlebins, bin_width = 1, bin_start = NULL)
}
\arguments{
\item{dat_for_mlebins}{tibble of data in the format required for fitting
using MLEbins method. TODO can join up with \code{determine_xmin_and_fit_mlebins()}}

\item{bin_width}{numeric bin width to fit a histogram to help determine xmin}

\item{bin_start}{numeric value for the first bin to start at; if \code{NULL} then
is set to the highest multiple of \code{bin_width} value below \code{min(dat_for_mlebins$bin_min)}.}
}
\value{
a histogram list object with components (see \code{?hist}):
\itemize{
\item \code{breaks}
\item \code{mids}
\item \code{counts}
\item \code{xname}
\item \code{equidist} TRUE since have equal bin widths
}
}
\description{
Called from \code{determine_xmin_and_fit_mlebins()} to determine xmin for MLEbins
method, and should work for MLEbin
method also TODO. Adapting from \code{sizeSpectra2::make_hist()} and
sizeSpectraHake::make_hist()`, think the first was adapted from the former,
for simple counts, but now we want the more complex latter version for
binned data.
}
\details{
Makes results a histogram list object and create 0 counts for missing bins.

Can then use \code{plot()} which calls \code{plot.histogram()}. Without the 0 counts for missing bins
\code{plot.histogram()} does not plot counts because bins appear to have unequal widths.
}
\examples{
\dontrun{
TODO counts_per_bin_example
make_hist(counts_per_bin_example)
TODO for test do
}
}
\author{
Andrew Edwards
}
