% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-for-mlebins.R
\name{mediterranean_for_mlebins}
\alias{mediterranean_for_mlebins}
\title{Extract and format the required Mediterranean data for MLEbins analysis, for
a given strata and group}
\usage{
mediterranean_for_mlebins(
  dat,
  group_name = NULL,
  strata_name = NULL,
  minimum_length = NULL,
  maximum_length = NULL
)
}
\arguments{
\item{dat}{tibble of data already with certain columns TODO}

\item{group_name}{group to analyse, if not specified then use all  TODO
think if can be vector, is.null might not work}

\item{strata_name}{strata to analyse, if not specified then use all}

\item{minimum_length}{remove fish shorter than this}

\item{maximum_length}{remove fish longer than this  TODO not incorporated yet}
}
\value{
tibble to go into \code{\link[=fit_size_spectrum]{fit_size_spectrum()}}, including the species
column so it uses MLEbins method TODO double check that flow through
}
\description{
See \code{zabala-data-analysis.Rmd}. This is specific to this data set, but can
easily be adapted/generalised for others
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
