% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.determine_xmin_and_fit.R
\name{plot.determine_xmin_and_fit}
\alias{plot.determine_xmin_and_fit}
\title{Plot results from determining xmin by mode method and then fitting use MLEhake spectra results}
\usage{
\method{plot}{determine_xmin_and_fit}(res, ...)
}
\arguments{
\item{res}{list of class \code{determine_xmin_and_fit} as output from \code{determine_xmin_and_fit()}}

\item{...}{arguments to passed onto TODO \code{sizeSpectra::ISD_bin_plot_nonoverlapping()}}
}
\value{
figure in current device
}
\description{
Plot results from determining xmin by mode method and then fitting use MLEhake spectra results
}
\examples{
\dontrun{
See .Rmd
}
}
\author{
Andrew Edwards
}
