% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.determine_xmin_and_fit_mlebins.R
\name{plot.determine_xmin_and_fit_mlebins}
\alias{plot.determine_xmin_and_fit_mlebins}
\title{Plot results from determining xmin by mode method and then fitting use
MLEbins method}
\usage{
\method{plot}{determine_xmin_and_fit_mlebins}(
  res,
  xlim_hist = NULL,
  par_mai = c(0.4, 0.5, 0.05, 0.3),
  par_cex = 0.7,
  seg_col = "green",
  ...
)
}
\arguments{
\item{res}{list of class \code{determine_xmin_for_mlebins_and_fit} as output from \code{determine_xmin_for_mlebins_and_fit()}}

\item{xlim_hist}{numeric vector of two values representing \code{xlim} for histogram plot; default is the full range
of breaks (which might be too large for a clear figure, especially given
the linear axis).}

\item{par_mai}{vector of values to use for \code{par(mai)}}

\item{par_cex}{numeric value to use for \code{par(cex)} (font size)}

\item{seg_col}{colour to use for the segments (green in Fig. 7 of
MEPS paper).}

\item{...}{arguments to passed onto \code{\link[=hist]{hist()}} or \code{\link[=plot.size_spectrum_mlebins]{plot.size_spectrum_mlebins()}}}
}
\value{
figure in current device  TODO inheritParams
plot.size_spectrum_mlebin   did not work TODO check everything flows through
}
\description{
Plot results from determining xmin by mode method and then fitting use
MLEbins method
}
\examples{
\dontrun{
See .Rmd
}
}
\author{
Andrew Edwards
}
