% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.determine_xmin_for_mlebins_and_fit.R
\name{plot.determine_xmin_and_fit_mlebins}
\alias{plot.determine_xmin_and_fit_mlebins}
\title{Plot results from determining xmin by mode method and then fitting use
MLEbins method # TODO function is actually the same as
plot.determine_xmin_and fit, since plot is already assigned correctly. So
might not need this one, or just call the other one. TODO TODO NOT quite,
two lines changed.}
\usage{
\method{plot}{determine_xmin_and_fit_mlebins}(res, xlim_hist = NULL, ...)
}
\arguments{
\item{res}{list of class \code{determine_xmin_for_mlebins_and_fit} as output from \code{determine_xmin_for_mlebins_and_fit()}}

\item{xlim_hist}{numeric vector of two values representing \code{xlim} for histogram plot; default is the full range
of breaks (which might be too large for a clear figure, especially given
the linear axis). TODO could maybe add in log-x axis, but not option in
\code{plot.hist()} so would have to make it up here.}

\item{...}{arguments to passed onto \code{\link[=hist]{hist()}} for now, but probably want
more. TODO \code{sizeSpectra::ISD_bin_plot_nonoverlapping()}}
}
\value{
figure in current device
}
\description{
Plot results from determining xmin by mode method and then fitting use
MLEbins method # TODO function is actually the same as
plot.determine_xmin_and fit, since plot is already assigned correctly. So
might not need this one, or just call the other one. TODO TODO NOT quite,
two lines changed.
}
\examples{
\dontrun{
See .Rmd
}
}
\author{
Andrew Edwards
}
