% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aggregate.R
\name{plot_aggregate}
\alias{plot_aggregate}
\title{Do an aggregated plot of several PLB fits, each to a separate (but related)
group of individuals.}
\usage{
plot_aggregate(
  res_list,
  col_vec = c("orange", "lightblue", "green", "magenta"),
  xlim_global = NULL,
  ylim_global = NULL
)
}
\arguments{
\item{res_list}{list of results, with each component a list object of a
given type TODO first doing for \code{size_spectrum_numeric}, might be somewhat
automatic to generalise. Am making plot_aggregate_mlebins() and prob
plot_aggregate_mlebin(); could check the class here and keep this as the
user outfacing function, or also make plot_aggregate_numeric and just call
the right one. Some of the details could be shared here maybe.}

\item{col_vec}{vector of colours to assign for each group}
}
\value{
nothing TODO, should this return something?
}
\description{
Given a list of MLE results (TODO check how many different types), combine
the data and show an aggregated distribtuion, as well as the individual fits.
}
\examples{
\dontrun{
# See aggregating-size-spectra.Rmd TODO copy something to here maybe
}

}
\author{
Andrew Edwards
}
