% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aggregate-fits.R
\name{plot_aggregate_fits}
\alias{plot_aggregate_fits}
\title{Plot the aggregate fits for several strata}
\usage{
plot_aggregate_fits(
  agg_list,
  strata_names,
  log_axes = "xy",
  restrict = TRUE,
  normalise = TRUE,
  col_strata = c("blue", "darkgreen", "red"),
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{agg_list}{list of aggregated fits, with each component corresponding
to the list object resulting from \code{\link[=plot_aggregate_mlebin]{plot_aggregate_mlebin()}} for a
particular strata.}

\item{strata_names}{vector of strings, the names to use to annotate each
curve in turn (names correspond to the order in \code{agg_list}).}

\item{log_axes}{which axes to log (gets passed to \code{plot(..., log = log_axes)}, either \code{"xy"} or \code{"x"} are of interest (although \code{"y"}
and \code{""} automatically work).}

\item{restrict}{logical, whether or not to restrict the aggregated fits to
only being above a common x value}

\item{normalise}{logical, whether or not to normalise each aggregated fit by
it's maximum, so that all fits are on the same scale}

\item{col_strata}{vector of strings, each being a colour to use for each strata}

\item{xlim, }{ylim vectors for axes limits; if NULL (the default) they get
calculated automatically, but users likely want to then manually adjust
them, especially \code{ylim} to be able to see all fits asymptoting at their
xmin values, but the
default axis will go way too low (it is hard to automate this, so just
refine \code{ylim} to celarly show the aymptotes).}

\item{...}{additional arguments passed onto \code{\link[=plot]{plot()}}}
}
\value{
list with two objects, \code{x_plb_agg} and \code{y_plb_agg}, which are the
fitted x and y values for the aggregated size spectrum (which does not
have a simple exponent). To then use for plotting multiple strata in \code{\link[=plot_aggregate_fits]{plot_aggregate_fits()}}.
}
\description{
Uses the results from \code{\link[=plot_aggregate_mlebin]{plot_aggregate_mlebin()}} to plot aggregated size
spectra for several strata on a single plot. Has options for normalising and
restricting x range to give justifiable comparisons.
}
\examples{
\dontrun{
# See aggregating-size-spectra.Rmd TODO copy something to here maybe
}

}
\author{
Andrew Edwards
}
