% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aggregate-mlebin.R
\name{plot_aggregate_mlebin}
\alias{plot_aggregate_mlebin}
\title{Do an aggregated MLEbin plot of several PLB fits, each to a separate (but related)
group of individuals.}
\usage{
plot_aggregate_mlebin(
  res_list,
  col_vec = c("orange", "lightblue", "green", "darkblue", "darkgreen"),
  col_agg = "magenta",
  xlim_global = NULL,
  ylim_global = NULL,
  y_scaling = 0.25,
  ...
)
}
\arguments{
\item{res_list}{list of results, with each component a list object of a
given type TODO share help with plot-aggregate.}

\item{col_vec}{vector of colours to assign for each group}
}
\value{
list with two objects, \code{x_plb_agg} and \code{y_plb_agg}, which are the
fitted x and y values for the aggregated size spectrum (which does not
have a simple exponent). To then use for plotting multiple strata in \code{\link[=plot_aggregate_fits]{plot_aggregate_fits()}}.
}
\description{
TODO forcussing on MLEbins first, but should be general enough but need
testing for MLEbin. This is copying \link{plot_aggregate} and then editing.
}
\details{
TODO Am making plot_aggregate_mlebin() and maybe
plot_aggregate_mlebins(), though prob best to make the latter the same,
and don't want the green lines as a bit too confusing maybe.; could check
the class here and have a general outfacing function,
or also make plot_aggregate_numeric and just call
the right one. Some of the details could be shared here maybe.

Given a list of MLEbin results, combine
the data and show an aggregated distribution, as well as the individual fits.
}
\examples{
\dontrun{
# See aggregating-size-spectra.Rmd TODO copy something to here maybe
}

}
\author{
Andrew Edwards
}
