% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-isd-binned.R
\name{plot_isd_binned}
\alias{plot_isd_binned}
\title{Plot a single binned ISD plot, as called from \code{plot.size_spectrum_mlebin()}}
\usage{
plot_isd_binned(
  res_mlebin,
  log,
  xlim,
  ylim,
  x_plb,
  y_plb,
  y_plb_conf_min,
  y_plb_conf_max,
  plot_conf_ints = TRUE,
  xlab = expression(paste("Values, ", italic(x))),
  ylab = expression(paste("Number of ", values >= x), sep = ""),
  mgp_val = c(1.6, 0.5, 0),
  tcl_small = -0.2,
  inset_label = c(0, 0),
  inset_text = c(0, 0.04),
  legend_label = NULL,
  legend_text = NULL,
  legend_text_n = NULL,
  legend_position = "topright",
  x_big_ticks = NULL,
  x_big_ticks_labels = NULL,
  x_small_ticks = NULL,
  x_small_ticks_by = NULL,
  x_small_ticks_labels = NULL,
  y_big_ticks = NULL,
  y_big_ticks_labels = NULL,
  y_small_ticks = NULL,
  y_small_ticks_by = NULL,
  y_small_ticks_labels = NULL,
  y_scaling = 0.75,
  par_mai = c(0.4, 0.5, 0.05, 0.3),
  par_cex = 0.7,
  seg_col = "green",
  rect_col = "grey",
  fit_col = "red",
  fit_lwd = 2,
  conf_lty = 2
)
}
\arguments{
\item{res_mlebin}{size_spectrum_mlebin object resulting from running
\code{fit_size_spectrum()} on binned data (such that the function
\code{fit_size_spectrum_mlebin()} is used; see vignette TODO). # TODO link help
to plot.size_spectrum_numeric - may want to put them all together}

\item{log}{Which axes to log, for \code{plot(..., log = log)}. So "xy" for
log-log axes, "x" for only x-axis logged.}

\item{x_plb}{vector of values to use to plot the fitted PLB curve; if NA then
automatically calculated (sometimes need to manually extend it to hit the
x-axis, but tricky to automate that on a log-scale}

\item{y_plb}{vector of values corresponding to the MLE fit of the PLB
distribution at each value of \code{x_PLB}}

\item{y_plb_conf_min, }{y_plb_conf_max vector of values corresponding to the fit, using the
minimum/maximum value confidence interval for exponent b, of the PLB
distribution at each value of \code{x_PLB}}

\item{plot_conf_ints}{logical whether to plot confidence intervals or not}

\item{xlab}{ylab x/y labels, explicitly given default values here which can
be modified as required.}

\item{mgp_val}{mgp values to use, as in \code{plot(..., mgp = mgp_vals)}; see \link{?par}.}

\item{tcl_small}{Length of small tickmarks; see \link{?axis}.}

\item{legend_text}{text to put in the corner, defaults to \verb{b = <value>}. Set to
\code{NA} to have none, and modify default if needed.}

\item{legend_position}{where to place legend, gets used as the first
argument in \code{\link[=legend]{legend()}}.}

\item{x_big_ticks}{y_big_ticks numeric vector of big tick marks on
x-axis/y-axis. If NULL then gets done automatically, so if that does not
look good then define explicitly here, plus \code{x_big_ticks_labels} and/or \code{y_big_ticks_labels}.}

\item{x_big_ticks_labels}{y_big_ticks_labels numeric vector of big tick
marks to label on x-axis/y-axis. If NULL then gets done automatically, so
if that does not look good then define explicitly here}

\item{x_small_ticks}{y_small_ticks  numeric vector of small tick marks on x-axis/y-axis. If
NULL then gets done automatically, so if that does not  look good then
define explicitly here.}

\item{x_small_ticks_by}{y_small_ticks_by  numeric vector of increment to
use to generate small tick marks on x-axis/y-axis. Will conincide with the
big tick marks and extend beyond them. Can only define \code{x_small_ticks} or
\code{x_small_ticks_by} (same for \code{y_...}). Only relevant for linear axes. Set
to \code{NA} to force no small unlabelled tickmarks.}

\item{x_small_ticks_labels}{y_small_ticks_labels numeric vector of big tick
marks to label on x-axis/y-axis. If NULL then gets done automatically, so
if that does not look good then define explicitly here}

\item{y_scaling}{numeric scaling of y-minimum of y-axis. Axis can't go to zero on
log-log plot, but goes to the proportion \code{y_scaling} (<1)
of the minimum value of counts greater than the highest \code{bin_min} value. Do
such that can see the right-most bin in all plots.}

\item{par_mai}{vector of values to use for \code{par(mai)}}

\item{par_cex}{numeric value to use for \code{par(cex)} (font size)}

\item{seg_col}{colour to use for the segments (green in Fig. 7 of
MEPS paper).}

\item{fit_col}{colour to use for fitted curves}

\item{fit_lwd}{line thickness to use for fitted curves}

\item{conf_lty}{line type to use for confidence intervals}
}
\value{
single panel plot of the ISD with data in binned form like in
Fig. 7a or 7b of MEPS paper, but with nonoverlapping bins; returns nothing.
}
\description{
Plot a single binned ISD plot, as called from \code{plot.size_spectrum_mlebin()}
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
